add flush mode to 1d fabric semaphore inc packets
defaults to true (flush outstanding writes) before writing out the semaphore inc. Users can override to no flush if they know there is no race.
