+++ tests/tt_metal/microbenchmarks/ethernet/test_fabric_edm_bandwidth.py
@@ -236,9 +236,13 @@ def run_fabric_edm(
     bw_threshold = 0.07
     if packet_size <= 2048 and fabric_mode != FabricTestMode.Linear:
         bw_threshold = 0.12
-    assert expected_bw - bw_threshold <= bandwidth <= expected_bw + bw_threshold, "Bandwidth mismatch"
+    assert (
+        expected_bw - bw_threshold <= bandwidth <= expected_bw + bw_threshold
+    ), f"Bandwidth mismatch. expected: {expected_bw} B/c, actual: {bandwidth} B/c"
     if expected_Mpps is not None:
-        assert expected_Mpps - 0.01 <= mega_packets_per_second <= expected_Mpps + 0.01, "Packets per second mismatch"
+        assert (
+            expected_Mpps - 0.01 <= mega_packets_per_second <= expected_Mpps + 0.01
+        ), f"Packets per second mismatch. expected: {expected_Mpps} Mpps, actual: {mega_packets_per_second} Mpps"
 
 
 @pytest.mark.parametrize("num_mcasts", [200000])
