+++ tt_metal/fabric/hw/inc/edm_fabric/fabric_edm_packet_transmission.hpp
@@ -1,4 +1,4 @@
-// SPDX-FileCopyrightText: © 2024 Tenstorrent Inc.
+// SPDX-FileCopyrightText: © 2025 Tenstorrent Inc.
 //
 // SPDX-License-Identifier: Apache-2.0
 
@@ -8,6 +8,7 @@
 #include "fabric_edm_packet_header.hpp"
 #include "edm_fabric_worker_adapters.hpp"
 #include "fabric_edm_types.hpp"
+#include "tt_metal/fabric/hw/inc/edm_fabric/1d_fabric_constants.hpp"
 #include <cstdint>
 
 // If the hop/distance counter equals to the below value, it indicates that it has
@@ -89,7 +90,10 @@ FORCE_INLINE void print_pkt_header(volatile tt::tt_fabric::LowLatencyPacketHeade
 // Since we unicast to local, we must omit the packet header
 // This function only does reads, and within scope there are no modifications to the packet header
 FORCE_INLINE void execute_chip_unicast_to_local_chip(
-    tt_l1_ptr PACKET_HEADER_TYPE* const packet_start, uint16_t payload_size_bytes, uint32_t transaction_id) {
+    tt_l1_ptr PACKET_HEADER_TYPE* const packet_start,
+    uint16_t payload_size_bytes,
+    uint32_t transaction_id,
+    uint8_t rx_channel_id) {
     const auto& header = *packet_start;
     uint32_t payload_start_address = reinterpret_cast<size_t>(packet_start) + sizeof(PACKET_HEADER_TYPE);
 
@@ -123,6 +127,14 @@ FORCE_INLINE void execute_chip_unicast_to_local_chip(
         case tt::tt_fabric::NocSendType::NOC_UNICAST_ATOMIC_INC: {
             const uint64_t dest_address = header.command_fields.unicast_seminc.noc_address;
             const auto increment = header.command_fields.unicast_seminc.val;
+            if (header.command_fields.unicast_seminc.flush) {
+                auto start_trid = RX_CH_TRID_STARTS[rx_channel_id];
+                auto end_trid = NUM_TRANSACTION_IDS;
+                for (size_t i = start_trid; i < end_trid; i++) {
+                    while (!ncrisc_noc_nonposted_write_with_transaction_id_flushed(
+                        tt::tt_fabric::edm_to_local_chip_noc, i));
+                }
+            }
             noc_semaphore_inc(dest_address, increment, tt::tt_fabric::edm_to_local_chip_noc);
 
         } break;
